refactor(render): relocate region_adapter + card bodies out of back/ (ADR-0038 Phase 2a)
Kills the render->back back-edges: moves the 10-module region_adapter/
subsystem + workspace_card_bodies.py into render/fragment/region/, and
extracts the pure _resolve_row_links helper into render/fragment/region/
_row_links.py (shared by back's standalone list path as back->render).
Rewrites ~22 import sites; back->render is the legal direction.

render/ now has zero back/ imports. Full non-e2e suite: 18047 passed.
ADR-0038 updated to record this continues #1086's layering migration.
