+++ src/dazzle/render/fragment/region/_builders_charts.py
@@ -28,10 +28,6 @@ from __future__ import annotations
 from html import escape as _html_escape
 from typing import Any, Literal
 
-from dazzle.back.runtime.renderers.region_adapter._shared import (
-    _region_title,
-    _wrap_surface,
-)
 from dazzle.render.fragment import (
     BarChart,
     BarTrack,
@@ -54,6 +50,10 @@ from dazzle.render.fragment import (
     Sparkline,
     Surface,
     TimeSeries,
+)
+from dazzle.render.fragment.region._shared import (
+    _region_title,
+    _wrap_surface,
 )
