+++ src/dazzle/render/fragment/region/_builders_metrics.py
@@ -17,10 +17,6 @@ from __future__ import annotations
 
 from typing import Any, Literal
 
-from dazzle.back.runtime.renderers.region_adapter._shared import (
-    _region_title,
-    _wrap_surface,
-)
 from dazzle.render.fragment import (
     EmptyState,
     Fragment,
@@ -32,6 +28,10 @@ from dazzle.render.fragment import (
     StatusList,
     StatusListEntry,
     Surface,
+)
+from dazzle.render.fragment.region._shared import (
+    _region_title,
+    _wrap_surface,
 )
