+++ src/dazzle/render/fragment/region/_builders_misc.py
@@ -19,11 +19,6 @@ from __future__ import annotations
 
 from typing import Any
 
-from dazzle.back.runtime.renderers.region_adapter._shared import (
-    _region_title,
-    _render_typed_value,
-    _wrap_surface,
-)
 from dazzle.render.fragment import (
     URL,
     ConfirmCheckItem,
@@ -37,6 +32,11 @@ from dazzle.render.fragment import (
     Surface,
     Tree,
     TreeNode,
+)
+from dazzle.render.fragment.region._shared import (
+    _region_title,
+    _render_typed_value,
+    _wrap_surface,
 )
 
 _LABEL_CANDIDATES: tuple[str, ...] = ("title", "name", "id")
