+++ tests/unit/test_dispatch_ctx_list_row_links.py
@@ -12,13 +12,11 @@ from __future__ import annotations
 
 import pytest
 
-from dazzle.back.runtime.renderers.fragment_adapter import (
-    FragmentSurfaceAdapter,
-    _resolve_row_links,
-)
+from dazzle.back.runtime.renderers.fragment_adapter import FragmentSurfaceAdapter
 from dazzle.core.ir.surfaces import SurfaceMode
 from dazzle.render.context import ColumnContext, TableContext
 from dazzle.render.fragment import FragmentRenderer, Table
+from dazzle.render.fragment.region._row_links import _resolve_row_links
 from dazzle.ui.runtime.page_routes import _build_dispatch_ctx
