+++ tests/unit/test_region_adapter.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from dazzle.back.runtime.renderers.region_adapter import WorkspaceRegionAdapter
 from dazzle.render.fragment import Region, Surface
+from dazzle.render.fragment.region import WorkspaceRegionAdapter
 from dazzle.render.fragment.renderer import FragmentRenderer
